refactor(global): #RBS-37 add folder to split file typescript + scss
wip
