+++ src/main/resources/public/css/additional.css
@@ -269,160 +269,3 @@ a.structure:hover {
         border-bottom: thin solid #fff !important;
     }
 }
-
-
-/* Evol additional style*/
-.input-size {
-    width: 40px;
-}
-
-.default-fontsize {
-    font-size: 16px !important;
-}
-
-.transparent {
-    opacity: 0;
-    height: 0;
-    width: 0;
-}
-
-.pointer {
-    cursor: pointer;
-}
-
-.blue {
-    color: blue;
-}
-
-i.md-icon {
-    font-size: 20px;
-}
-
-i.suppr::before {
-    @include fonticon();
-    content: '\e8cf';
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    background-image: none;
-    margin-right: 5px;
-    color: dimgrey;
-    vertical-align: middle;
-}
-
-i.modif::before {
-@include fonticon();
-    content: '\e858';
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    background-image: none;
-    margin-right: 5px;
-    color: dimgrey;
-    vertical-align: middle;
-}
-
-i.valid::before {
-@include fonticon();
-    content: '\e871';
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    background-image: none;
-    margin-right: 5px;
-    color: dimgrey;
-    vertical-align: middle;
-}
-
-i.cancel::before {
-@include fonticon();
-    content: '\e866';
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    background-image: none;
-    margin-right: 5px;
-    color: dimgrey;
-    vertical-align: middle;
-}
-
-i.date::before {
-    @include fonticon();
-    content: '\e917';
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    background-image: none;
-    margin-right: 5px;
-    color: dimgrey;
-    vertical-align: middle;
-}
-
-i.add::before {
-    @include fonticon();
-    content: '\e85e';
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    background-image: none;
-    margin-right: 5px;
-    color: dimgrey;
-    vertical-align: middle;
-}
-
-.classic-button {
-    background: none !important;
-    box-shadow: none;
-    padding: 0 !important;
-    margin: 0 !important;
-}
-
-.resource-table {
-    table-layout:fixed;
-    width: 100%;
-    margin-top: 1rem;
-}
-
-.resource-table tr td {
-    padding: 0.5rem 0;
-}
-
-.resource-table tr td {
-    text-align: center;
-}
-
-#addUnavailability td input {
-    margin: 0;
-}
-
-.selected {
-    color: darkorange;
-    font-weight: bold;
-}
-
-.check-boxes {
-    padding: 0;
-    margin: 0;
-    clear: both;
-}
-
-.check-boxes li {
-    list-style-type: none;
-    list-style-position: outside;
-    float: left;
-    padding: 0;
-}
-
-.check-boxes input[type="checkbox"]:not(:checked),
-.check-boxes input[type="checkbox"]:checked {
-    position: absolute;
-    left: -9999%;
-    opacity: 0;
-    height: 0;
-    width: 0;
-}
-
-.check-boxes input[type="checkbox"] + label {
-    cursor: pointer;
-    color: #555;
-}
-
-.check-boxes input[type="checkbox"]:checked + label {
-    cursor: pointer;
-    color: darkorange;
-    font-weight: bold;
-}
