+++ src/main/resources/public/sass/global/components/_calendar.scss
@@ -1,0 +1,12 @@
+calendar {
+  switch {
+    .switch {
+      .tick {
+        &:before {
+          height: 20px !important;
+          margin-top: 0 !important;
+        }
+      }
+    }
+  }
+}
