+++ src/main/resources/public/sass/global/components/_index.scss
@@ -1,0 +1,3 @@
+@import "icons";
+@import "check-boxes";
+@import "resource-table";
