+++ src/main/resources/public/ts/app.ts
@@ -1,6 +1,6 @@
 import {ng} from 'entcore';
-import {RbsController} from './controller';
-import {timePicker} from './additional';
+import {RbsController} from './controllers/controller';
+import {timePicker} from './directives/additional';
 
 ng.controllers.push(RbsController);
 ng.directives.push(timePicker);
