+++ src/main/resources/view-src/admin.html
@@ -7,7 +7,7 @@
     <script src="/rbs/public/dist/entcore/ng-app.js" id="context" type="text/javascript"></script>
     <script src="/rbs/public/dist/application.js" type="text/javascript"></script>
 </head>
-<body ng-controller="RbsController">
+<body ng-controller="RbsController" class="rbs">
 <admin-portal>
     <h1>RBS</h1>
