+++ src/main/resources/view-src/rbs.html
@@ -12,7 +12,7 @@
     <link href="/rbs/public/css/timepicker.css" rel="stylesheet">
     <script type="text/ng-template" id="empty"></script>
 </head>
-<body ng-controller="RbsController">
+<body ng-controller="RbsController" class="rbs">
 <portal>
     <div class="row">
         <container template="top-menu"></container>
